Final 1.14 Commit
I wanted to commit this before I port the current project to 1.15

+++ src/main/java/com/builderboy/elementis/common/item/StaffItem.java
@@ -59,7 +59,7 @@ public class StaffItem extends ManaContainerItem {
         int current = nbt.getInt("mana");
         int max = type.calculateStoredMana();
 
-        return 1 - (float)(current / max);
+        return 1 - (double)current / (double)max;
     }
 
     @Override
